requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
pymysql
pydantic[email]
requests
python-dotenv
boto3
python-multipart
pillow